% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrapped_timeSeries}
\alias{bootstrapped_timeSeries}
\title{Obtain bootstrap replicate of time series}
\usage{
bootstrapped_timeSeries(cpts, x)
}
\description{
Obtain bootstrap replicate of time series
}
\keyword{internal}
