% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comb_contains_cpt}
\alias{comb_contains_cpt}
\title{Does the combination comb of changepoints contain the changepoint k_ind?}
\usage{
comb_contains_cpt(comb, k_ind)
}
\description{
Does the combination comb of changepoints contain the changepoint k_ind?
}
\keyword{internal}
