% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpts_bootstrap_help}
\alias{cpts_bootstrap_help}
\title{Helping function to get bootstrap replicates of change point estimates}
\usage{
cpts_bootstrap_help(cpts_info, x, N_reps)
}
\description{
Helping function to get bootstrap replicates of change point estimates
}
\keyword{internal}
