% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleMosum_merging.R
\name{dup.merge}
\alias{dup.merge}
\title{Remove duplicated from all.cpts data frame:
In case one change being added multiple times, choose the one
with smallest p-Value (resp. highest peak)}
\usage{
dup.merge(all.cpts, rule = "peak")
}
\description{
Remove duplicated from all.cpts data frame:
In case one change being added multiple times, choose the one
with smallest p-Value (resp. highest peak)
}
\keyword{internal}
