% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_criterion_help}
\alias{eta_criterion_help}
\title{extract changepoints from candidates with eta criterion}
\usage{
eta_criterion_help(candidates, m_values, eta, G_left, G_right)
}
\description{
extract changepoints from candidates with eta criterion
}
\keyword{internal}
