% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_sub}
\alias{extract_sub}
\title{Helping function for algorithm 2: Pre-compute the partial sums
S_i = sum{j=k_i+1}^{k_{i+1}}x_i and the partial sums of squared
T_i = sum{j=k_i+1}^{k_{i+1}}x_i^2
between the (sorted) candidates k_i and k_{i+1} in cand.
Output: data frame with 4 columns k_i | k_{i+1} | S_i | T_i}
\usage{
extract_sub(cand, x)
}
\description{
Helping function for algorithm 2: Pre-compute the partial sums
S_i = sum{j=k_i+1}^{k_{i+1}}x_i and the partial sums of squared
T_i = sum{j=k_i+1}^{k_{i+1}}x_i^2
between the (sorted) candidates k_i and k_{i+1} in cand.
Output: data frame with 4 columns k_i | k_{i+1} | S_i | T_i
}
\keyword{internal}
