% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_comb_ind}
\alias{get_comb_ind}
\title{Get integer vector of changepoint indices,
based on bool-field representation of combinations.
E.g. for combination index 11 [=1011]: get_comb_ind(c(T,F,T,T))=11}
\usage{
get_comb_ind(active)
}
\description{
Get integer vector of changepoint indices,
based on bool-field representation of combinations.
E.g. for combination index 11 [=1011]: get_comb_ind(c(T,F,T,T))=11
}
\keyword{internal}
