% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_k_star}
\alias{get_k_star}
\title{Compute bootstrapped mosum statistic and return maximum position thereof}
\usage{
get_k_star(x_star, k_hat, G_l, G_r)
}
\description{
Compute bootstrapped mosum statistic and return maximum position thereof
}
\keyword{internal}
