% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_local_costs}
\alias{get_local_costs}
\title{Compute the Local cost terms of combination icomb (for RSS resp. sBIC).
Use pre-computed partial sum matrix sub_sums (see extract_sub) for speedup}
\usage{
get_local_costs(icomb, sub_sums)
}
\description{
Compute the Local cost terms of combination icomb (for RSS resp. sBIC).
Use pre-computed partial sum matrix sub_sums (see extract_sub) for speedup
}
\keyword{internal}
