% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prune.R
\name{local.env}
\alias{local.env}
\title{Identify the local environment for exhaustive search}
\usage{
local.env(j, est.cpts.ind, all.cpts, current, ac)
}
\description{
Identify the local environment for exhaustive search
}
\keyword{internal}
