% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prune.R
\name{local.prune}
\alias{local.prune}
\title{Localised pruning algorithm}
\usage{
local.prune(x, all.cpts, rule, log.penalty, pen.exp)
}
\description{
Localised pruning algorithm
}
\keyword{internal}
