% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_help}
\alias{mean_help}
\title{helping function for bootstrap (compute local means)}
\usage{
mean_help(x, l, r)
}
\description{
helping function for bootstrap (compute local means)
}
\keyword{internal}
