% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewiseStationary_modelSignals.R
\name{modelSignal.blocks}
\alias{modelSignal.blocks}
\title{Creating Time-Series according to example model}
\usage{
modelSignal.blocks(rand.gen = rnorm, ...)
}
\arguments{
\item{rand.gen}{optional: a function to generate the innovations}

\item{...}{further arguments to be parsed to \code{rand.gen}}
}
\value{
a vector consisting of a realization of the signal model
}
\description{
Creates a time series according to the block model with independent innovations.
}
\details{
See Appendix B in the reference for details about the signal model.
}
\references{
P. Fryzlewicz.
\emph{Wild Binary Segmentation for Multiple Change-Point Detection.}
The Annals of Statistics, Vol. 42, No. 6, 2243-2281, 2014.
}
\keyword{internal}
