% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_models.R
\name{modelSignal.mix}
\alias{modelSignal.mix}
\title{Creating Time-Series according to example model}
\usage{
modelSignal.mix(rand.gen = rnorm, ...)
}
\arguments{
\item{rand.gen}{optional: a function to generate the innovations}

\item{...}{further arguments to be parsed to \code{rand.gen}}
}
\value{
a vector consisting of a realization of the signal model
}
\description{
Creates a time series according to the block model with independent innovations.
}
\details{
See Appendix B in the reference for details about the signal model.
}
\references{
P. Fryzlewicz (2014)
Wild Binary Segmentation for Multiple Change-Point Detection.
\emph{The Annals of Statistics}, Volume 42, Number 6, pp. 2243-2281.
}
\keyword{internal}
