% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mosum.cpts.bootstrap}
\alias{mosum.cpts.bootstrap}
\title{Bootstrapping MOSUM changepointss}
\usage{
mosum.cpts.bootstrap(mcpts, N_reps, alpha)
}
\description{
Bootstrapping MOSUM changepointss
}
\keyword{internal}
