% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_test.R
\name{mosum.pValue}
\alias{mosum.pValue}
\title{MOSUM asymptotic p-value}
\usage{
mosum.pValue(z, n, G.left, G.right = G.left)
}
\arguments{
\item{z}{a numeric value for the observation}

\item{n}{an integer value for the length of the input data}

\item{G.left, G.right}{integer values for the left moving sum bandwidth (G.left,G.right)}
}
\value{
a numeric value for the asymptotic p-value for the asymmetric MOSUM test
}
\description{
Computes the asymptotic p-value for the MOSUM test.
}
\keyword{internal}
