% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_test.R
\name{mosumCriticalValue}
\alias{mosumCriticalValue}
\title{MOSUM asymptotic critical value}
\usage{
mosumCriticalValue(n, G.left, G.right, alpha)
}
\arguments{
\item{n}{an integer value for the length of the input data}

\item{G.left, G.right}{integer values for the left moving sum bandwidth (G.left,G.right)}

\item{alpha}{a numeric value for the significance level with
\code{0 <= alpha <= 1}}
}
\value{
a numeric value for the asymptotic critical value for the MOSUM test
}
\description{
Computes the asymptotic critical value for the MOSUM test.
}
