% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{multiscale.cpts.bootstrap}
\alias{multiscale.cpts.bootstrap}
\title{Bootstrapping multiscale MOSUM changepoints}
\usage{
multiscale.cpts.bootstrap(ms, N_reps, alpha)
}
\description{
Bootstrapping multiscale MOSUM changepoints
}
\keyword{internal}
