% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{next_bit_permutation}
\alias{next_bit_permutation}
\title{Next value to iterate (in lexicographical order) over all bit 
permutaions having l bits set to 1.
Example sequence (2 bits): {0011, 0101, 0110, 1001, 1010, 1100}.
Source: https://stackoverflow.com/questions/1851134/generate-all-binary-strings-of-length-n-with-k-bits-set}
\usage{
next_bit_permutation(v)
}
\description{
Next value to iterate (in lexicographical order) over all bit 
permutaions having l bits set to 1.
Example sequence (2 bits): {0011, 0101, 0110, 1001, 1010, 1100}.
Source: https://stackoverflow.com/questions/1851134/generate-all-binary-strings-of-length-n-with-k-bits-set
}
\keyword{internal}
