% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numberOfSetBits}
\alias{numberOfSetBits}
\title{Get number of non-zero bits of a 32bit integer
Source: https://stackoverflow.com/questions/109023/how-to-count-the-number-of-set-bits-in-a-32-bit-integer}
\usage{
numberOfSetBits(i)
}
\description{
Get number of non-zero bits of a 32bit integer
Source: https://stackoverflow.com/questions/109023/how-to-count-the-number-of-set-bits-in-a-32-bit-integer
}
\keyword{internal}
