% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum.R
\name{plot.mosum}
\alias{plot.mosum}
\title{Plotting MOSUM objects}
\usage{
\method{plot}{mosum}(x, alpha = 0.05, critical.value.col = "blue",
  xlab = "Time", ...)
}
\arguments{
\item{x}{a \code{mosum} object}

\item{alpha}{a numeric value for the significance level with
\code{0 <= alpha <= 1}}

\item{critical.value.col}{a specification for the color of the
critical value, see \link[graphics]{par}}

\item{xlab}{graphical parameter}

\item{...}{additional graphical arguments, see \link[graphics]{plot}}
}
\description{
Plotting method for objects of class "mosum".
}
