% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleMosum_merging.R
\name{plot.multiscale.cpts}
\alias{plot.multiscale.cpts}
\title{Plotting multiscale.cpts changepoints}
\usage{
\method{plot}{multiscale.cpts}(x, shaded = c("CI", "bandwidth")[1],
  N_reps = 10000, alpha_CI = 0.05, include_uniform_CI = F, ...)
}
\arguments{
\item{x}{an \code{multiscale.cpts} object}

\item{shaded}{string indicating what to display as shaded area on the y-axis around change point estimates.
Poissble values are \code{"bandwidth"} for the respective detection bandwidth and \code{"CI"} for (bootstrapped) confidence intervals;}

\item{N_reps, alpha_CI}{argument to be parsed to \code{cpts.bootstrap}. Use iff \code{shaded="CI"}.}

\item{include_uniform_CI}{shall uniform confidence intervals also be plotted? Use iff \code{shaded="CI"}.}

\item{...}{not in use}
}
\description{
Plotting method for objects of class "multiscale.cpts".
}
