% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum.R
\name{print.mosum.cpts}
\alias{print.mosum.cpts}
\title{Change-points estimated by MOSUM procedure}
\usage{
\method{print}{mosum.cpts}(x, ...)
}
\arguments{
\item{x}{a \code{mosum.cpts} object}

\item{...}{not in use}
}
\description{
Print method for objects of class \code{mosum.cpts}
}
\examples{
x <- testData(lengths = rep(100, 3), means = c(0, 5, -2), sds = rep(1, 3), seed = 1234)$x
m <- mosum(x, G = 40)
print(m)
}
