% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prune.R
\name{print.multiscale.cpts}
\alias{print.multiscale.cpts}
\title{Change-points estimated by multiscale MOSUM procedure}
\usage{
\method{print}{multiscale.cpts}(x, ...)
}
\arguments{
\item{x}{a \code{multiscale.cpts} object}

\item{...}{not in use}
}
\description{
Print method for objects of class \code{multiscale.cpts}
}
\examples{
x <- testData(model = "mix", seed = 12345)$x
mlp <- multiscale.localPrune(x)
print(mlp)
}
