% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rolling_sum}
\alias{rolling_sum}
\title{equivalent to zoo::rollsum(x, k=G, fill=NA, align="left"), 
but optimized for speed}
\usage{
rolling_sum(x, G)
}
\description{
equivalent to zoo::rollsum(x, k=G, fill=NA, align="left"), 
but optimized for speed
}
\keyword{internal}
