% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setBitNumber}
\alias{setBitNumber}
\title{where is leftmost one?
https://www.geeksforgeeks.org/find-significant-set-bit-number/}
\usage{
setBitNumber(n)
}
\description{
where is leftmost one?
https://www.geeksforgeeks.org/find-significant-set-bit-number/
}
\keyword{internal}
