% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{start_bit_permutations}
\alias{start_bit_permutations}
\title{Starting value to iterate (in lexicographical order) 
over all bit permutaions having l bits set to 1.
E.g.: start_bit_permutations(2) = 3 [=0..011].}
\usage{
start_bit_permutations(l)
}
\description{
Starting value to iterate (in lexicographical order) 
over all bit permutaions having l bits set to 1.
E.g.: start_bit_permutations(2) = 3 [=0..011].
}
\keyword{internal}
