% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{cpts_bootstrap}
\alias{cpts_bootstrap}
\title{Helping/wrapper fuction for C++ calls}
\usage{
cpts_bootstrap(cpts_info, N_reps, level, mcpts)
}
\description{
Helping/wrapper fuction for C++ calls
}
\keyword{internal}
