% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prune.R
\name{dup.merge}
\alias{dup.merge}
\title{Remove duplicated from all.cpts data frame:
In case one change being added multiple times, choose the one
with smallest p-value}
\usage{
dup.merge(all.cpts)
}
\description{
Remove duplicated from all.cpts data frame:
In case one change being added multiple times, choose the one
with smallest p-value
}
\keyword{internal}
