% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum.R
\name{plot.mosum.cpts}
\alias{plot.mosum.cpts}
\title{Plotting MOSUM change-points}
\usage{
\method{plot}{mosum.cpts}(x, critical.value.col = "blue", cpts.col = "red",
  xlab = "Time", ...)
}
\arguments{
\item{x}{a \code{mosum.cpts} object}

\item{critical.value.col}{a specification for the color of the
critical value, see \link[graphics]{par}}

\item{cpts.col}{a specification for the color of the
change-points, see \link[graphics]{par}}

\item{xlab}{graphical parameter}

\item{...}{additional graphical arguments, see \link[graphics]{plot}
and \link[graphics]{abline}}
}
\description{
Plotting method for objects of class \code{mosum.cpts}
}
\details{
The input time series is plotted along with vertical lines indicating
the estimated change-point locations and a horizontal line indicating
the critical value used in the MOSUM test.
}
